Student contribution keyboard for Indoeuropean langauges written with the Latin alphabet. From the submission e-mail:
«Hello,

I made a keyboard for the roman letters of the indo-european languages (well, superficially). Would be great if it were added to the site!


Best regards,
Jarno Porkka»

Dir structure etc provided by template, layout by student. Submitted as part of the course «Linguistics in the Digital Age» at University of Helsinki.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/startup-keyboards/ine@160198 c7155fb1-f0a7-4240-a2fc-2600b6f42f90